7.4 Разделит редьюсер по зонам ответственности: состояние игры, загружаемые данные.
