+++ src/const.js
@@ -3,11 +3,4 @@ export const GameType = {
   GENRE: `genre`,
 };
 
-export const ActionType = {
-  INCREMENT_MISTAKES: `INCREMENT_MISTAKES`,
-  INCREMENT_STEP: `INCREMENT_STEP`,
-  RESET_GAME: `RESET_GAME`,
-  LOAD_QUESTIONS: `LOAD_QUESTIONS`,
-};
-
 export const MAX_MISTAKES = 3;
