+++ src/reducer/name-space.js
@@ -1,0 +1,4 @@
+export const NameSpace = {
+  DATA: `DATA`,
+  GAME: `GAME`,
+};
