--- src/reducers/selectors.js
@@ -1,11 +1,0 @@
-export const getQuestions = (state) => {
-  return state.steps.questions;
-};
-
-export const getSteps = (state) => {
-  return state.steps.step;
-};
-
-export const getMistakes = (state) => {
-  return state.mistakes;
-};
